refactor user service to interface level:
