+++ src/main/java/com/rova/accountservice/controller/UserController.java
@@ -4,8 +4,8 @@ import com.rova.accountservice.dals.IUserDetails;
 import com.rova.accountservice.dto.FUserDto;
 import com.rova.accountservice.dto.ResponseDto;
 import com.rova.accountservice.exceptions.CommonsException;
-import com.rova.accountservice.exceptions.UserNotFoundException;
 import com.rova.accountservice.services.impl.UserService;
+import com.rova.accountservice.services.impl.UserServiceImpl;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.Authorization;
 import io.swagger.annotations.AuthorizationScope;
